% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvimclip.R
\name{nvimclip}
\alias{nvimclip}
\title{nvimclip}
\usage{
nvimclip(objname = "mtcars")
}
\arguments{
\item{objname}{Name of the object to be processed. Default is "mtcars".}
}
\value{
JSON representation of the processed object, copied to the clipboard.
}
\description{
nvimclip
}
