% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvimclip.R
\name{process_character}
\alias{process_character}
\title{process_character}
\usage{
process_character(x, name)
}
\arguments{
\item{x}{Character object to be processed.}

\item{name}{Name of the object.}
}
\value{
Processed character content as a string.
}
\description{
process_character
}
