% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvimclip.R
\name{process_contents}
\alias{process_contents}
\title{process_contents}
\usage{
process_contents(x, name)
}
\arguments{
\item{x}{Object to be processed.}

\item{name}{Name of the object.}
}
\value{
Processed content as a string.
}
\description{
process_contents
}
