% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvimclip.R
\name{process_else}
\alias{process_else}
\title{process_else}
\usage{
process_else(x, name)
}
\arguments{
\item{x}{Object to be processed.}

\item{name}{Name of the object.}
}
\value{
Processed content as a string for objects that are not numeric, character, factor, or logical.
}
\description{
process_else
}
