% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvimclip.R
\name{process_numeric}
\alias{process_numeric}
\title{process_numeric}
\usage{
process_numeric(x, name)
}
\arguments{
\item{x}{Numeric object to be processed.}

\item{name}{Name of the object.}
}
\value{
Processed numeric content as a string.
}
\description{
process_numeric
}
